pool_partner_price_service/pyproject.toml
[tool.poetry]
name = "pool_partner_price_service"
version = "0.1.0"
description = ""
authors = ["Bernardo Gomes <bgomesdeabreu@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.7"
prettyconf = "^2.2.1"
kafka-python = "^2.0.2"
requests = "^2.26.0"
psycopg2 = "^2.9.1"
SQLAlchemy = "^1.4.25"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
